feat: move files
resolves: #23
