+++ src/main/kotlin/com/kroffle/knitting/controller/design/DesignRouter.kt
@@ -1,6 +1,6 @@
-package com.kroffle.knitting.presentation.router.design
+package com.kroffle.knitting.controller.design
 
-import com.kroffle.knitting.domain.handler.design.DesignHandler
+import com.kroffle.knitting.usecase.design.DesignHandler
 import org.springframework.context.annotation.Bean
 import org.springframework.context.annotation.Configuration
 import org.springframework.web.reactive.function.server.RequestPredicates.path
