+++ src/main/kotlin/com/kroffle/knitting/infra/design/DesignRepository.kt
@@ -1,6 +1,5 @@
-package com.kroffle.knitting.data.repository.design
+package com.kroffle.knitting.infra.design
 
-import com.kroffle.knitting.data.entity.design.DesignEntity
 import org.springframework.data.repository.reactive.ReactiveCrudRepository
 
 interface DesignRepository : ReactiveCrudRepository<DesignEntity, Long>
