+++ src/main/kotlin/com/kroffle/knitting/infra/design/DesignType.kt
@@ -1,4 +1,4 @@
-package com.kroffle.knitting.data.entity.design
+package com.kroffle.knitting.infra.design
 
 enum class DesignType(val code: Int) {
     Sweater(1),
