+++ src/main/kotlin/com/kroffle/knitting/usecase/ping/PingHandler.kt
@@ -1,4 +1,4 @@
-package com.kroffle.knitting.domain.handler.ping
+package com.kroffle.knitting.usecase.ping
 
 import org.springframework.http.MediaType
 import org.springframework.stereotype.Component
